Fix : group_rouine_final_2

+++ server/app/controllers/group_routines.js
@@ -221,7 +221,7 @@ module.exports = {
   },
     
 
-// GET : localhost:4000/group-routine/select?id=3&data=1646715662218
+// GET : localhost:4000/group-routine/select?id=3&date=1646715662218
   select_group_routine: {
     get: async (req, res) => {
       // 나의 토큰 이메일과, 그룹루틴 id를 이용하여 그룹루틴 클릭하기.
@@ -244,7 +244,6 @@ module.exports = {
         // localhost:4000/group-routine/select?id=2&date=1646715662218
         const groupRoutineID = +req.query.id;
         const date = +req.query.date;
-        console.log('date', date)
 
         // 유저테이블에서 나의 user.id 찾기
         const myUserId = await user
@@ -284,9 +283,8 @@ module.exports = {
           return result;
         };
 
-        // xxxx-xx-xx 형식의 선택된 데이터값
+        // xxxx-xx-xx 형식의 선택된 데이터 (년-월-일)
         const selectedDate = `${year()}-${month()}-${day()}`;
-        console.log('selectedDate -------------------------- ', selectedDate)
 
         // 해당 날짜의 댓글 가져오기
         const selectedComment = await comment.findAll({
@@ -298,36 +296,45 @@ module.exports = {
           }, raw: true
         });
 
+        const fillterdselectedComment = []
+        const filterComment = () => {
+          for(let el of selectedComment) {
+            if(!fillterdselectedComment.includes(el)) {
+              fillterdselectedComment.push(el);
+            }
+          }
+        }
+        await filterComment();
+
+
         // 달성률 체크 : 선택된 커맨트 갯수 / 가입자수
-        const goal =  Math.floor((selectedComment.length / joinedMember.length) * 100)
-        console.log('goal', goal) 
+        const goal =  Math.floor((fillterdselectedComment.length / joinedMember.length) * 100)
+        console.log('goal', goal)
 
         // 유저테이블에서 comment 작성자 이름 찾기
         const findUserName = async (userid) => {
           const userName = await user
           .findOne({ raw: true, where: { id: userid } })
           .then(user => user.nickname)
-          // .then((data) =>  console.log('data --------------- ', data.nickname))
-          // console.log('userName', userName)
           return userName
         }
 
-        // 해당 날짜의 댓글 보내기 => [{writer: nickname, comment: '댓글입니다.'}, {writer: nickname2, comment: '댓글입니다2.'}, ... ]
+
+        // TODO: 유저네임, 타임 같이 넣어주기, 댓글 똑같은 사람이 여러개 달았을떄 문제가있음, 해당 날짜 추가된 댓글을 comments 안에 배열로 보내기
+
+        // 해당 날짜의 댓글 보내기 => [{writer: nickname, comment: '댓글입니다.', time : '15:23'}, {writer: nickname2, comment: '댓글입니다2.', time : '05:25'}, ... ]
         // 그냥 하면 맵을 써서 맵안에서 findOne하면 펜딩이 되는데 다음과 같이 Array를 따로 만들어 줘서하면 pending 안됨.. 이유는??
         let selectedCommentArray = [];
         const makeCommentArray = async () => {
           for(let el of selectedComment) {
             const name = await findUserName(el.user_id)
-            selectedCommentArray.push({ writer: name, comment: el.comment})
-            // console.log('selectedCommentArray', selectedCommentArray)
+            selectedCommentArray.push({ writer: name, comment: el.comment, time: `${el.createdAt.getHours()}:${el.createdAt.getMinutes()}`});
           }
         }
         await makeCommentArray();
 
         // editor 여부 판단하기 editor: true or false;
         const imEditor = selectedGroupRoutineData.editor_id === myUserId ? true : false
-        console.log('imEditor', imEditor)
-
 
         if (myGroupId.includes(groupRoutineID)) {
           // 가입된 그룹 루틴 선택 : 코멘트를 쓰는 인풋창 있음, 콘텐츠, 날짜선택, 댓글, 루틴달성률, 그룹탈퇴하기 버튼
@@ -339,8 +346,7 @@ module.exports = {
             editor: imEditor,
             registed: true,
             message: "가입한 그룹 루틴 데이터",
-          });
-          // return res.json('응 가입한 그룹맞아')
+          })
         } else {
           // 가입안된 그룹 루틴 선택 : 콘텐츠, 날짜선택, 댓글, 루틴달성률, 그룹가입하기 버튼, 전체 멤버수
           // TODO: 전체 멤버수, 해당 날 달성률, 댓글작성자 이름 (comments : [{writer: overflowbin, commemt: '안녕하세요~~'}])
@@ -357,10 +363,8 @@ module.exports = {
   },
 
 
-
-
-  // 댓글 남기기
-
+  // FIXME: 댓글 남기기
+  // POST : localhost:4000/group-routine/comment?id=2&date=1646900498082
   write_comment: {
     post: async (req, res) => {
 
@@ -382,7 +386,7 @@ module.exports = {
       const groupRoutineID = +req.query.id;
 
       // 요청 받은 코멘트 
-      // const { comment } = await req.body;
+      // const { comment } = await req.body; -> comment table이 있어 변수가 겹친다. 쓰면 안됨.
 
       // 유저테이블에서 나의 user.id 찾기
       const myUserId = await user
@@ -391,8 +395,68 @@ module.exports = {
 
       // 받은 코멘트 데이터 베이스에 넣기 -> 날짜는 저절로 바뀜
       // 근데 만약 수정을 한다? ===> 우선 수정, 삭제 키 없이 하기로!
-      await comment.create({ comment: req.body.comment, user_id: myUserId, group_routine_id: groupRoutineID })
-      return res.status(201).json({ data: null, message: "댓글이 추가되었습니다" })
+      await comment.create({ 
+        comment: req.body.comment, 
+        user_id: myUserId, 
+        group_routine_id: groupRoutineID })
+
+
+      // 추가된 댓글 + 해당 루틴, 해당 날에 있는 댓글 다시 보내기
+      const date = +req.query.date;
+
+      // 선택된 달, 요일
+      const year = () => new Date(date).getFullYear();
+      const month = () => {
+        let result = new Date(date).getMonth() + 1;
+        if (result < 10) {
+          return (result = "0" + result);
+        }
+        return result;
+      };
+      const day = () => {
+        let result = new Date(date).getDate();
+        if (result < 10) {
+          return (result = "0" + result);
+        }
+        return result;
+      };
+
+      // xxxx-xx-xx 형식의 선택된 데이터 (년-월-일)
+      const selectedDate = `${year()}-${month()}-${day()}`;
+
+      // 해당 날짜의 댓글 가져오기
+      const selectedComment = await comment.findAll({
+        where: {
+          [Op.and]: [
+            { group_routine_id: groupRoutineID },
+            { createdAt: { [Op.startsWith]: selectedDate } },
+          ],
+        }, raw: true
+      });
+
+      // 유저테이블에서 comment 작성자 이름 찾기
+      const findUserName = async (userid) => {
+        const userName = await user
+        .findOne({ raw: true, where: { id: userid } })
+        .then(user => user.nickname)
+        return userName
+      }
+
+      // 해당 날짜의 댓글 보내기 => [{writer: nickname, comment: '댓글입니다.', time : '15:23'}, {writer: nickname2, comment: '댓글입니다2.', time : '05:25'}, ... ]
+      // 그냥 하면 맵을 써서 맵안에서 findOne하면 펜딩이 되는데 다음과 같이 Array를 따로 만들어 줘서하면 pending 안됨.. 이유는??
+      let selectedCommentArray = [];
+        const makeCommentArray = async () => {
+          for(let el of selectedComment) {
+            const name = await findUserName(el.user_id)
+            selectedCommentArray.push({ 
+              writer: name, 
+              comment: el.comment, 
+              time: `${el.createdAt.getHours()}:${el.createdAt.getMinutes()}`});
+          }
+        }
+        await makeCommentArray();
+
+      return res.status(201).json({ data: { comment : selectedCommentArray }, message: "댓글이 추가되었습니다" })
     },
   },
 
@@ -400,23 +464,60 @@ module.exports = {
   // 루틴 그룹 삭제하기
   // 삭제에 대한 여부 파악은 client에서 조건부 랜더링으로 파악된 상황임.
   // 쿼리로 해당 그룹 루틴 id 받아오기
-  // destroy
   delete_group_routine: {
     get: async (req, res) => {
 
-      // 그룹아이디
-      const groupRoutineID = +req.query.id;
+      // 토큰을 통해 유저 아이디 받아오기
+      function getCookie(name) {
+        let matches = req.headers.cookie.match(
+          new RegExp(
+            "(?:^|; )" +
+              name.replace(/([\.$?*|{}\(\)\[\]\\\/\+^])/g, "\\$1") +
+              "=([^;]*)"
+          )
+        );
+        return matches ? decodeURIComponent(matches[1]) : undefined;
+      }
+      const accessToken = getCookie("accessToken");
+      const { email } = jwt.verify(accessToken, process.env.ACCESS_SECRET);
+
+      // 유저테이블에서 나의 user.id 찾기
+      const myUserId = await user
+      .findOne({ where: { email } })
+      .then((data) => data.id);
+
 
       try {
 
-        // 해당 댓글 삭제하기
-        // 해당그룹에 소속된 모든 댓글 삭제하기
-        await comment.destroy({ where: { group_routine_id: groupRoutineID }})
+        // 그룹 루틴 아이디
+      const groupRoutineID = +req.query.id;
 
-        // 그룹루틴 삭제하기
-        await group_routine.destroy({ where: { id: groupRoutineID }});
+        // 삭제후 또 삭제하려는것을 방지 하기 위해 삭제 여부 확인 => 현재 시스템에서는 작동안됨.
+        const deleteRoutine = await group_routine.findOne({ where: { id: groupRoutineID }})
 
-        return res.status(200).json({ data: null, message: "그룹 루틴이 삭제되었습니다"});
+        // 로그인 한 사람이 해당그룹의 editor인지 확인
+        // editor 여부 판단하기 editor: true or false;
+        const selectedGroupRoutineData = await group_routine.findOne({
+          where: { id: groupRoutineID }, raw: true
+        });
+        const imEditor = selectedGroupRoutineData.editor_id === myUserId ? true : false
+
+        if(!imEditor) {
+          return res.status(203).json({ data: null, message: "삭제할 권한이 없습니다"});
+        } else {
+          if(!deleteRoutine) {
+            return res.status(203).json({ data: null, message: "이미 삭제된 그룹루틴입니다"});
+          } else {
+            // 해당 댓글 삭제하기
+            // 해당그룹에 소속된 모든 댓글 삭제하기
+            await comment.destroy({ where: { group_routine_id: groupRoutineID }})
+  
+            // 그룹루틴 삭제하기
+            await group_routine.destroy({ where: { id: groupRoutineID }});
+  
+            return res.status(200).json({ data: null, message: "그룹 루틴이 삭제되었습니다"});
+          }
+        }        
       }
 
       catch {
